refactor: common 모듈/패키지로 설정 파일 이동 및 의존성 정리
