+++ server/src/main/java/com/ahmadda/common/config/JpaAuditingConfig.java
@@ -1,4 +1,4 @@
-package com.ahmadda.domain.config;
+package com.ahmadda.common.config;
 
 import org.springframework.context.annotation.Configuration;
 import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
