+++ server/src/main/java/com/ahmadda/common/logging/LoggingDiscardPolicy.java
@@ -1,4 +1,4 @@
-package com.ahmadda.infra.logger;
+package com.ahmadda.common.logging;
 
 import lombok.extern.slf4j.Slf4j;
