--- server/src/main/java/com/ahmadda/infra/database/ReplicationRoutingDataSource.java
@@ -1,15 +1,0 @@
-package com.ahmadda.infra.database;
-
-import org.springframework.jdbc.datasource.lookup.AbstractRoutingDataSource;
-import org.springframework.transaction.support.TransactionSynchronizationManager;
-
-public class ReplicationRoutingDataSource extends AbstractRoutingDataSource {
-
-    @Override
-    protected Object determineCurrentLookupKey() {
-        if (TransactionSynchronizationManager.isCurrentTransactionReadOnly()) {
-            return "reader";
-        }
-        return "writer";
-    }
-}
